Finished CLevelManger Implementation
Game now cycles through MainMenu, CMLCentralCavern, and CManicLayer (infinitely).

Program crashes on second level after player collides with an invader, the crash happens during the VOnReset, specifically on one of the Platforms's Reset function.
This is probably because the ObjGroups are being duplicated, and when the CGCObjectManager runs OnReset(), it accesses corrupt data.