Implemented Umeer's CAirManager
Worked with Umeer to implement his CAirManager class.

It will only work for the first level, as some of its UI Sprites are being autoreleased when the level transition is over. We really don't know why, as we make sure to remove it from the parent level ( CMLCentralCavern ), but the autorelease gets triggered anyways when CMLCentralCavern calls VOnResourceReleased, which goes through its children and deletes them.

Calling retain() won't make a difference, as VOnResourceReleased calls the children's autorelease() instead of release().