Full Screen & Exit Button
Added Fullscreen functionality to the game layer, however the resolution remains the same. Also updated the Graphics in menu with something a bit more modern and an exit button was added to the menuscene and gamelayer, as exiting the game can only be done via that exit button in fullscreen.