Follow up on Zaf's feedback 13/10/2020
Changed CGameInstance and CLevelManager so they are no longer singletons.

Removed CGameInstance from the build, as it is ultimately pointless now.
Moved all the logic out of CGameInstance into CLevelManager (some of it is not functional yet such as collisions, they're commented out for now).

Massively improved program flow, CLevelManager is now newed in AppDelegate, and deleted when the program closes. CLevelManager creates the MainMenu scene instead of the other way around now.

Refactored the code that deals with the player going into the next level ( much easier to read, edit and expand ).

Added comments.