Started working towards Audio Functionality
Added ManicAudio.h, which is a helper header file:

This header file includes a const static std::map that links ESoundNames with their respective audio file directory.
It also includes a couple of functions to load and play sound effects, that take in an ESoundName, so we can easily play sound from anywhere just by including this header file.

The header file will work with the cocos2d SimpleAudioEngine.