Fixed 2 ( and hopefully all ) Crashes
Crash #1:
Fixed a crash that was happening when the Layer was being deleted after the transition ended, inside CManicLayer::VOnDestroy() an object group was not being unregistered and another one was being unregistered but not actually deleted.
These 2 things were causing not only memory leaks but horrible mess-ups inside the nodes's linked lists.
This crash was happening when trying to release the layer, it would crash inside the Unlink() function of a certain node, as its "NextNode" was nullptr.

Crash #2:
Fixed a crash that happened during the VOnReset() of whatever second level was loaded. The VOnResourceReleased() of the first level was deleting the sharedGB2ShapeCache, effectively deleting the physics shapes that were loaded during the creation of the second level. With this, the AddFixturesToBody() which is called during VOnReset() would crash, as there was no physics shapes in cached memory.
Fixed it by commenting out the line of code that reset the shape cache, as it really didn't need to get released in the first place, it's always the same.