Implemented jumping through platforms
The only platforms that work for now are the ground and the 1x1 individual blocks, as I only used those 2 assets to test on a whitebox level.
For the rest, I only need to assign sensors using the same methodology, through the PhysicsEditor, very easily done.

Only 1 issue needs fixing now, which is when walking on individual blocks in a row, sometimes a new block will stop you from moving that way. Even though the blocks in a row have the same y coordinate, I suspect the collision between the new block and the player is acting out, because no code seems to be originating this problem. It feels like the block is simply 1 pixel above its neighbour, so it stops you from moving.