Door class, Door creator added in
3 classes are added.
CDoor:: Which is the door itself, will handle it's own interact event.

CDoorCreator:: Superclass to be inherited from by the DoorCreators for each level, it requires a reference to Cmaniclayer as it will feed it down to the CDoor.

CDCreatorCentralCavern:: Inherits from CDoorCreator, it creates and destructs the CDoor for the first level.