Fixed issue with player
When standing under a brick platform, if player jumped it would cause a bug and the player wouldn't be able to move ever again.

This was fixed with a raycast implementation, when player tries to jump it now checks if it can actually jump before doing so. If the path directly above its head is blocked then it simply won't jump.