CGameManager implementation complete
A lot of things got changed.
SLevelValues > Struct that feeds to the GameManager to know the win condition for the level.
This is set on each individual level.
ERequiredType > Essentially a rework of the ECollectibleTypeRequired, except it doesn't have the ::Switch as it is unused. This enum is used in the SLevelValues.
CGameManager > The Highscore value is now stored in Highscore.bin. It is then encrypted by the program on storing the highscore, which a tempering check on start up to see if everything is as it should be.
Aside from that, it now controls from a high level the gamestates, this allows for the Air to be drained into the score.
Furthermore, the extra lives do work. Every time you hit 10k points, you get one.
CAirManager > For ease of use, added a bool with a getter for it for the CGameManager to use. This allows to control the drain and gamestate.
I have also added a pointer to the GameManager, which is set in the level Init. This is to use Umeer's once per frame drain (other wise I would call it every update which leads ofc to int overflow). This is not nice and I will eventually adapt his calculation into a function inside the CGameManager and remove the pointer inside AirManager.