Fixed issue with player sliding after jumping
Fixed the second issue Zaf mentioned during the tutorial. As I had predicted, the collision + physics engine was causing it outside my code.

The physics shape defined in PhysicsEditor was 1 pixel off in one of the vertices, which made it have a slight slope, which caused the tiny sideways acceleration.