Fixed Player Movement Bugs
Fixed the issue where player was getting stuck when moving along 1x1 blocks. This was being caused by the physics engine in reality, because player's Y coordinate would be something like 3.9998753 instead of 4, so sometimes he would be able to walk freely and sometimes he would bump against the next platform in front of him (completely random, not easily replicated).

Fixed this issue by manually adjusting the position to something like 4.00100 everytime the player lands, disabling the gravity, and only reenabling it when player jumps or drops from ledge.

Fixed another issue that was making the player start dropping from ledge when in fact he was standing on a block.