All level classes implemented
All other 16 classes have been made for the levels, set up with correct properties which means all it requires is the OGMO levels themselves.

I've added a new param to SLevelCreationParams > szLevelName. This will be displayed on top where the score is, like the original game. This has been implemented for every level as well.

Pls dont over ride the filter if merge issue.