Fixed bugs + 19_TheFinalBarrierLayout
Essentially in manic miner, some doors are actually not on a platform which means as it used to be, the following scenario would have happened:
Jump to door > air drains to score infinitely > player falls

So the work around is:
Jump to door > air drains until it reaches 0 > asks for next level.

With the way I did this, the player model gets hiden while drained (I have to check Manic miner to see what happens when you finish a level  whos door floating.
An alternative work around would be to disable gravity on the player when he hits the door.

Additionally, I have made the 19_TheFinalBarrier layout.