Fix to AnimationStateMachine on GameEnd
Needed to check whether player was alive, before changing states and playing animations, because ApplyDirectionChange is also called when the player has died 3 times and the main menu is loaded back in.  The player was already dead and the game was trying to play an animation, that's what caused the crash.