Fixed Player Jump Bug
When standing very near to a ledge and trying to jump sideways, player would always jump vertically. This was happening because on LeftGround() I was checking if the soft sensors were equal to 0 to drop straight down.

The fix was just to change the logic check. The previous check was ( no soft contacts ) OR ( can jump ), changed it to ( no soft contacts ) AND  ( can jump ).

Check CPlayer.cpp line 799