Changed Conveyor Belt Logic
I noticed the original game's conveyor belt mechanic worked a bit differently than what I had initially thought.
I fixed the way that the jumps worked, player is now free to jump the way he's walking, and not necessarily jump in the direction of the conveyor belt ( this is how it was working, even if the player was walking against that direction );

Fixed the landing / walking against the conveyor belt. Previous to this commit, it was just checking where the player wanted to go upon landing, but I've changed this to match the original, and now it checks not only that, but also makes sure the player was headed that way when initiating the jump prior to this landing.
So if player wants to walk left when the conveyor belt is directed to the right, the player must jump from the right to the left, and keep holding the left arrow when landing.

Had to add a member variable ( EPlayerDirection m_eJumpDirection ) to keep track of this extra info that needs checking.