Scrapped old implementation
The old implementation I had in mind wouldn't work well with the Parallax System, so the AnimatedParallaxLayer approach will be used instead, as it's already part of the parallax system.

Also added a couple of functions for getting random points in 2d within a certain range and calculating distances, which the "ai" of the fish will use to determine when to start moving towards the next random point.

As to how it'll work:
- Generate Random point
- start moving towards it
- once it has travelled 2/3 of the distance
- find next random point
- start interpolating/moving towards that point
- repeat entire process