Fixed conveyor belt bug
Sometimes after dropping from a ledge of the conveyor belt, the player would still be able to start moving in that direction and effectively get stuck moving in that direction in midair.

This was fixed by removing the SensorContactCount check when leaving the ground, to check if the player has jumped or is dropping straight down.