Movement Test & Calculation Functions
A simple movement test with moveto, to find out how moveto works. (The current location of the sprite needs to be taken into consideration, because that is used as the initial location for moveto, so 0,0 will be the sprite's current location in 2d space). Also added a function to get 2/3 of the distance value, that will be used to determine when to start moving towards the next point.