Jump Animation Integration & Fish Moving
Added Jump Animation and transition back to idle/run animation from tje jump animation, based on EPlayerDirection when the player lands.

Fish now moves right/left and then switches to moving in the opposite direction, when it reaches a certain point. And Fish now resets properly with the correct Vec2 Position, when the player dies. Before it got reset to 0,0, causing it to not show up anymore.