Fixed a bug on CTriggerPlatform
This class was causing a crash when VOnResourceRelease() was called, as it was releasing the resources (sprite) and then doing a check that involved that sprite afterwards. The fix was to just call the parent version of VOnResourceRelease() on the last line of this class's VOnResourceRelease(). As any should really.