Updated and finalized the Solar Light mechanic
Had to change a bit of logic inside the CManicLayer and CMLSolarPowerGenerator level classes to expose the AirManager. Now the CSolarLight has access to the air manager, and keeps updating it, telling it if the player is or is not overlapping with light.
This being done, the air is being drained 5x quicker than normal when player is standing on the light.
This feature is officially complete.
Could use some further refactoring in the future though.