Fixed a bug with climbing up the ledge mechanic
This mechanic is present only on brick platforms, if the player jumps and bumps against the wall, he will keep going up. If the jump is high enough to reach past the top surface of this brick wall, the player will automatically start moving forward again (in mid air). This mechanic is present in the original game, although there is no special animation for it. It is clearly visible in levels such as Eugene's Lair, in spots where you have to jump above 2 block tall brick walls, like the door spot at the bottom of this level.

The bug that was present was the player would just keep moving forward after landing on the top surface, this is because I forgot to update the movement properly using ApplyDirectionChange( ), instead I was just setting the instant velocity directly on the player.