Finished fixing platform collision bug
The bug: player would sometimes reach a state where falling had no movement lock, and when dropping from ledges the player would be able to move freely in midair.
Reason for the bug: player's m_iHardContactCount was being incremented but not decremented occasionally for some contacts. This resulted in being impossible to reach m_iHardContactCount == 0 ever again on a certain level, and the player wouldn't ever truly "leave the ground".

One of the sources of this bug was the crumbling platform. When it was done crumbling, it wasn't setting its "bCollisionEnabled" and "TriggersHardContacts" back to false.

Added a property to the parent platform class "m_bIsTriggeringHardContact", with a getter and a setter. This is updated inside the CManicLayer code, right after we trigger the HardContactEvent function on the player. An alternative to this would be, instead of having integers in the player class to keep track of how many contacts it has in any given frame, perhaps have a list of entities it's currently colliding with. This way, we wouldn't need the "IsTriggeringHardContact" inside each entitity, we just needed to look and see if said entity was listed on the "CurrentContacts" list the player would have. I thought this more advanced system wasn't necessary for a game such as this one but perhaps it would have been a good approach after all. Won't be changing it now, as that's a huge change to my player movement / collision system.

The bug was also happening partially because we were forcing a ghosted hard contact, and for doing so we were checking if the player had 0 hard contacts. This has been now changed, to not check the number of current contacts, because that's essentially a wrong approach. We are instead checking if this specific platform is being overlapped, but has not triggered a hard contact yet, and if it hasn't, it means this contact should now be forced, hence the name ghost contact.

It's also checking if the player is not moving upwards, because if this triggered while the player was jumping through a platform, it would make the player land on the platform instantly when reaching the ground level, the player would be grounded for 1 frame, and then would leave the ground again, as the vertical speed was still higher than 0. Essentially you could see that the jump animation flickered for 2 frames, so this way we avoid it.

Some other lines of code were also moved around, but just so that this new implementation would work correctly with all platforms.

Also slightly changed the way animations are played, so now the "Jumping" animation will be played when player leaves the ground. When dropping from a ledge, or falling from a crumbling platform that has crumbled away, the shark will throw his arms up in excitement, instead of standing still.