Fixed conveyor belt bug
Added a "m_bIsOnConveyorBelt" property for the player.
This is set when entering / leaving a hard contact with a conveyor belt.

Edited the code inside OnLanded( ) to check m_bIsOnConveyorBelt, and added a function for when the player leaves the conveyor belt.

These changes in logic allow the player to walk on top of a surface and then begin conveyor belt movement seamlessly, as well as going against it. It was acting very strangely with multiple different crazy outcomes, so that bug has been fixed this way.