Implemented footsteps sound
When moving, footsteps sound will start playing.
Also fixed a minor bug with audio, the audio clips were not stopping when transitioning to next level, so I created the onExitTransitionDidStart( ) override on CManicLayer, that calls the player's stop audio functions. The audios also stop when OnEscape( ) is called, in other words, when player walks through the portal door.