Animation State Machine Refactoring
iCounter was being initialized twice inside the LoadAnimations Function, so I removed the iCounter variable from inside (if(bShouldLoadAnimations == true)), as we already have needed initialized variable at the top of this function. Also moved the Dictionary Creation out of the for loop, as it needs to be created only once.